src/test/javascript/spec/app/entities/company-info/company-info-update.component.spec.ts
/* tslint:disable max-line-length */
import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { Observable, of } from 'rxjs';

import { WorkshopTestModule } from '../../../test.module';
import { CompanyInfoUpdateComponent } from 'app/entities/company-info/company-info-update.component';
import { CompanyInfoService } from 'app/entities/company-info/company-info.service';
import { CompanyInfo } from 'app/shared/model/company-info.model';

describe('Component Tests', () => {
    describe('CompanyInfo Management Update Component', () => {
        let comp: CompanyInfoUpdateComponent;
        let fixture: ComponentFixture<CompanyInfoUpdateComponent>;
        let service: CompanyInfoService;

        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [WorkshopTestModule],
                declarations: [CompanyInfoUpdateComponent]
            })
                .overrideTemplate(CompanyInfoUpdateComponent, '')
                .compileComponents();

            fixture = TestBed.createComponent(CompanyInfoUpdateComponent);
            comp = fixture.componentInstance;
            service = fixture.debugElement.injector.get(CompanyInfoService);
        });

        describe('save', () => {
            it(
                'Should call update service on save for existing entity',
                fakeAsync(() => {
                    // GIVEN
                    const entity = new CompanyInfo(123);
                    spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
                    comp.companyInfo = entity;
                    // WHEN
                    comp.save();
                    tick(); // simulate async

                    // THEN
                    expect(service.update).toHaveBeenCalledWith(entity);
                    expect(comp.isSaving).toEqual(false);
                })
            );

            it(
                'Should call create service on save for new entity',
                fakeAsync(() => {
                    // GIVEN
                    const entity = new CompanyInfo();
                    spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
                    comp.companyInfo = entity;
                    // WHEN
                    comp.save();
                    tick(); // simulate async

                    // THEN
                    expect(service.create).toHaveBeenCalledWith(entity);
                    expect(comp.isSaving).toEqual(false);
                })
            );
        });
    });
});
